refactor(work): move the work CLI adapter into the Work service transports tree
Relocate pkg/transports/cli/work to
pkg/services/work/transports/cli/work so the Work CLI adapter lives beside
the service it serves. Its only non-test service import is pkg/services/work.

Pure relocation: all 14 source files move unchanged and the 15 external
importers are updated to the new import path. Import blocks are re-sorted by
gofmt, which also settles the climanifest import left unsorted by #2057.
