+++ pkg/services/work/transports/cli/work/list_counts_json_test.go
