+++ pkg/services/work/transports/cli/work/list_filters_test.go
