+++ pkg/services/work/transports/cli/work/output_required_test.go
