+++ pkg/services/work/transports/cli/work/show.go
