+++ pkg/services/work/transports/cli/work/show_test.go
