+++ pkg/services/work/transports/cli/work/visualize.go
