+++ pkg/transports/cli/commandregistry/work_handlers.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"io"
 
+	workcli "github.com/portpowered/infinite-you/pkg/services/work/transports/cli/work"
 	"github.com/portpowered/infinite-you/pkg/transports/cli/resolvedinput"
-	workcli "github.com/portpowered/infinite-you/pkg/transports/cli/work"
 	"github.com/spf13/cobra"
 )
