+++ pkg/transports/cli/root_run_flags_test.go
@@ -8,9 +8,9 @@ import (
 	"fmt"
 	startupcli "github.com/portpowered/infinite-you/pkg/initializer/process"
 	"github.com/portpowered/infinite-you/pkg/services/work"
+	workcli "github.com/portpowered/infinite-you/pkg/services/work/transports/cli/work"
 	factorycli "github.com/portpowered/infinite-you/pkg/transports/cli/factory"
 	runcli "github.com/portpowered/infinite-you/pkg/transports/cli/run"
-	workcli "github.com/portpowered/infinite-you/pkg/transports/cli/work"
 	factoryapi "github.com/portpowered/infinite-you/pkg/transports/http/generated"
 	"github.com/spf13/cobra"
 	"io"
